fixes quicksurf/quicksurf-bash-utilities#3
The bug resulted in an internal variabled being named the same as the
script name so every time the script tried to reference the internal
variable, it ended up calling itself. This was resolved by renaming the
internal variable to $branch_name.

During testing of this script it was also discovered that the script
needed to check to see if the private branch it was switching to
already existed, and if so, act accordingly. This resulted in a new
script (be, meaning branch_exists) that checks to see if a branch
exists and if so, return success, otherwise false.

	new file:   be
	modified:   pb